refactor(orders_page): renaming presentation Layer
